RTC Story 36901 - Use LPC Memory
This includes a hack to allow access to our fake PNOR data via
the ECCB scom registers.  This hack will be removed once Simics
provides a real ECCB model.

Changes to INTR testcase were needed due to bugs exposed by the
timing changes when enabling this new code.

Note that the default operating mode will remain LPC_MEM because
the current version of the ECCB model causes the IPL to take
close to 10 minutes to complete.

Change-Id: Icc236bffd52ba8214ec920f9a496adec138e54d9
Reviewed-on: http://gfw160.austin.ibm.com:8080/gerrit/692
Tested-by: Jenkins Server
Reviewed-by: Daniel M. Crowell <dcrowell@us.ibm.com>