Per Mike -- adding back the removed chips.
In the weekly "scrum of scrums" meeting this morning, we touched on your change to add DIMMs to the targeting XML files.
I asked you to reduce the number of chips from 8 to 4 because there are no systems planned with more than 4 chips in a
drawer.  However, Dan Crowell said that he still wanted 8 chips because that is what the hardware supports. Patrick agreed.

So... Would you mind adding back those 4 chips?
I'm thinking that you can start with the old XML file and just use your script to add in the 512 DIMMs. I don't think this
is a big change.  Apologies for the confusion.

Change-Id: I17e172eb48007dd56f24c8a530a3ea1ef6ac62a0
Reviewed-on: http://gfw160.austin.ibm.com:8080/gerrit/678
Tested-by: Jenkins Server
Reviewed-by: MIKE J. JONES <mjjones@us.ibm.com>
Reviewed-by: Daniel M. Crowell <dcrowell@us.ibm.com>
Reviewed-by: A. Patrick Williams III <iawillia@us.ibm.com>