Create attribute dump-to-trace function
Work for Task 38385
I modified the attribute parsing script (xmltohb.pl) to generate
a debug function that will loop over all targets and trace out
the value of every defined attribute.  Note that this is not the
permanent direction, eventually we will do the parsing offline
using the debug framework.  This is a stop-gap until that is
ready.

For review I have a copy of the trace output here:
/gsa/ausgsa/home/d/c/dcrowell/public/hostboot/RTC_35202_output.txt

RTC: 35202

Change-Id: I038dfd0fd0eb10b25203fe91fa23933fd24fe62f
Reviewed-on: http://gfw160.austin.ibm.com:8080/gerrit/770
Tested-by: Jenkins Server
Reviewed-by: Andrew J. Geissler <andrewg@us.ibm.com>
Reviewed-by: Nicholas E. Bofferding <bofferdn@us.ibm.com>
Reviewed-by: A. Patrick Williams III <iawillia@us.ibm.com>