HWPF: Infrastructure to allow HWPs to look for specific PLAT errors
Currently when PLAT firmware (e.g. FSP, Hostboot, Cronus) attaches
an error log to a fapi::ReturnCode, the ReturnCode value is
FAPI_RC_PLAT_ERR_SEE_DATA, all a HWP receiving the ReturnCode knows
is that there is a PLAT error and it cannot look at the data (it is
PLAT specific). There will be future scenarios where HWPs need to
look for specific errors code from PLAT. The example given by Joe
McGill is ADU_LOCKED, a HWP may want to retry a SCOM operation if
the error is ADU_LOCKED. The way to do this will be to define a
PLAT return code value in fapiReturnCodes.H called
FAPI_RC_PLAT_ADU_ERR. PLAT code would set this special value and
attach its error log. Then HWP code can look for this special
value. This change is simply to allow PLAT code to specify the
return code value when attaching an error log to the ReturnCode.

Change-Id: I6bc3f0b4d3350cc3eed52fa947a5bf4a885095a5
RTC: 37970
Reviewed-on: http://gfw160.austin.ibm.com:8080/gerrit/785
Tested-by: Jenkins Server
Reviewed-by: Brian H. Horton <brianh@linux.ibm.com>
Reviewed-by: A. Patrick Williams III <iawillia@us.ibm.com>