HWPF: Hostboot errlogs should cope with HWP error IDs greater than 255
RTC: 39175

This is **part** of the work required for RTC 39175.
When a HWP creates an error, the reason code is
(HWPF_COMP_ID | errID). The errID is a single byte. Error log
tags are automatically created so that the errors can be
parsed. However at some point there will be more than 255
HWP errors so this scheme will not work. And more pressing is
RTC 39175, the error IDs will change from 0, 1, 2, 3, to
32-bit hash codes. The new plan is to have a fixed reason
code for HWP errors (HWPF_COMP_ID | RC_HWP_GENERATED_ERROR),
the 32 bit error ID will be FFDC data which will be parsed
by an auto-generated function. This also changes the expected
HWPF unit test error to be committed by CXXTEST which is the
subject of a request by Dan.

Change-Id: I5669f2a53702649d4599556a9e380ec3725ca51b
Reviewed-on: http://gfw160.austin.ibm.com:8080/gerrit/792
Tested-by: Jenkins Server
Reviewed-by: Brian H. Horton <brianh@linux.ibm.com>
Reviewed-by: A. Patrick Williams III <iawillia@us.ibm.com>