fix prcd_server.tcl and prcd_compile.tcl to better detect/handle errors.
user reported an issue w/ the tool, where the build wasn't happening, but
the prcd_compile.tcl script reported no error. investigation showed that
the prcd_server.tcl script was not able to create the sandbox directory
in /tmp because the OS had removed the /tmp/hwp directory. fixes:
 . server script will make the /tmp/hwp directory each extract time
 . if server script has an error, the msg back to the compile script
    will start ERROR:
 . compile script will look for ERROR: messages and report them to the
    user as a server script error and exit w/ an error code.

Change-Id: I90dee6d57e807fc274e0a44210412d7007c199b0
RTC: 39966
Reviewed-on: http://gfw160.austin.ibm.com:8080/gerrit/821
Tested-by: Jenkins Server
Reviewed-by: MIKE J. JONES <mjjones@us.ibm.com>
Reviewed-by: A. Patrick Williams III <iawillia@us.ibm.com>