Improve presence detection in HWAS code
Change presence  detection logic in HWAS code to be a bit more general, in
case something in the layout and structure of CLASS and TYPE changes. Instead
of relying on a CLASS_SYS -> CLASS_ENC -> (objects that need physical detect)
hierarchy, rely on the fact that we know all PROC MEMBUF and DIMM targets are
the ones that need the platform-specific physical detect.

Also, list of present targets will be culled so that only CLASS_CHIP targets
are in the list for the call to the platform-specific routine that does
the Chip ID and EC Level fetching.

And, use new HUID attribute instead of local function for debug/error msgs.

This is RTC Task 39791.

Change-Id: I3f3e14ebcc02da15ab08a2c8d9151dbff6648a34
RTC: 35007
Reviewed-on: http://gfw160.austin.ibm.com:8080/gerrit/836
Reviewed-by: A. Patrick Williams III <iawillia@us.ibm.com>
Tested-by: Jenkins Server