Cause build to fail if duplicate reason code / module id is found
Change the scanforsrcs.pl script to die w/ an error if a duplicate
reason code is found within a module. In addition, correct makefile
so that the script is run all the time, so that errors are caught
before jenkins. And more script to code_pass instead of gen_pass
so that all files it needs are created before it's run.

Change-Id: Idd5fe546207920e9561091bbed1817efc04f2ebc
RTC: 38731
Reviewed-on: http://gfw160.austin.ibm.com:8080/gerrit/869
Tested-by: Jenkins Server
Reviewed-by: Terry J. Opie <opiet@us.ibm.com>
Reviewed-by: A. Patrick Williams III <iawillia@us.ibm.com>