fix bugs in prcd_compile tool
update prcd_server.tcl to output better error messages, and messages that
the prcd_compile script will receive and handle properly. update the list
of files that are returned on a build to be all that are in the cpfiles.pl
script. also, clarify help/usage in prcd_compile.tcl output

defect SW135271 "prcd_compile.tcl fails to create the host boot image" was
opened for this.

Change-Id: Ia283c4c54cabd80bc8c4c460f3ef987c37b00a06
Reviewed-on: http://gfw160.austin.ibm.com:8080/gerrit/904
Tested-by: Jenkins Server
Reviewed-by: MIKE J. JONES <mjjones@us.ibm.com>
Reviewed-by: A. Patrick Williams III <iawillia@us.ibm.com>