Errlog User Details: Log Arbitrary Attributes
Add code so that a hostboot or procedure developer can save attributes
into an error log and have them display in a human-readable way in order
to aid debug and FA.  All data is stored in errlog in binary format and
decoded into strings and enum 'constants' with the parse function.

Methods to dump and parse the attributes are auto-generated by the
xmltohb.pl script.

In addition, cleanup up Target User Details so that blanks lines aren't
output and so that the attribute name isn't put into the error log and
is instead added by the parse function (smaller errlog usage), and change
to same coding style as other UserDetail functions (data put into buffer
in ctor and not in addToLog function).

following will be added in future story/task work:
 - parsing for complex attributes.
 - combining multiple attribute user details together instead of
    generating a separate log entry for each.

Change-Id: I6baafa36ef095aafabb5bf86d6054bcb3db47019
RTC: 35275
Reviewed-on: http://gfw160.austin.ibm.com:8080/gerrit/944
Reviewed-by: A. Patrick Williams III <iawillia@us.ibm.com>
Tested-by: Jenkins Server