platReadIDEC killing IPL for single failure
Correct HWAS code - if target is detected as present, but the read
of the ID/EC fails, then mark the target as present but NOT
functional, commit the errl, and continue the target processing.
All physical children under that parent target will be marked as
present and NOT functional as well.

Same for the deviceRead for the Presence detect - if it fails,
mark the target as not present, commit the errl, and continue.

Change-Id: I6d911c6e6aa3968395721dc2826bf860589dc737
RTC: 41123
Reviewed-on: http://gfw160.austin.ibm.com:8080/gerrit/974
Tested-by: Jenkins Server
Reviewed-by: A. Patrick Williams III <iawillia@us.ibm.com>