Disable PNOR Flat Model testing.
Since no one is actually using the temporary "FLAT" PNOR model in
simics, I told the simics team the could default to the 'REAL'
pnor model from now on.  This was causing a test case to fail,
so removing it for now.  Story 35728 will add support for REAL
PNOR and corresponding test cases.

Change-Id: Ifb976dea227b4e996fedd37fb762ae1c5411aad8
Reviewed-on: http://gfw160.austin.ibm.com:8080/gerrit/975
Tested-by: Jenkins Server
Reviewed-by: Daniel M. Crowell <dcrowell@us.ibm.com>
Reviewed-by: A. Patrick Williams III <iawillia@us.ibm.com>