HWAS code common between hostboot and FSP
Corrections to common and hostboot platform-specific code to make this
code work for both hostboot and FSP. In addition, make printing of HUID
consistant with other previous changes.

Also, added hwas/common files to fsp distribution tools so that they are
included in the fsp.tar for common usage.

Change-Id: Ia761606ab446dd794bf8e112430cc2e072f46991
RTC: 39793
Reviewed-on: http://gfw160.austin.ibm.com:8080/gerrit/1130
Tested-by: Jenkins Server
Reviewed-by: A. Patrick Williams III <iawillia@us.ibm.com>