make prcd_compile and prcd_server more robust and user-friendly.
add catch calls to prcd_server so that if the user ctrl-C's out of the
client, we have less of a risk of things hanging. add more output
to prcd_comiple so that the user gets more of an indication that things
are happening, so that they hopefully won't ctrl-C out in the middle.

Change-Id: Iecb656f029bd7b849212751be320be2d0b7ceea5
RTC: 42908
Reviewed-on: http://gfw160.austin.ibm.com:8080/gerrit/1133
Tested-by: Jenkins Server
Reviewed-by: A. Patrick Williams III <iawillia@us.ibm.com>