Change PNORDD test cases to use test section of PNOR
Updated the PNOR Device Driver test cases to use a specially
created test section of PNOR.  The test cases will only run if
the TEST section exists.

Disabled the fake-pnor related test cases as the test section
offset does not exist in the fake-pnor address space.  Opened
a story to fix this later.

Change-Id: I1b5fd4989ee775c14034430226d9ffe844995f96
Reviewed-on: http://gfw160.austin.ibm.com:8080/gerrit/1335
Tested-by: Jenkins Server
Reviewed-by: A. Patrick Williams III <iawillia@us.ibm.com>