HWPF: Minor fapi::Target fix and FAPI tracing tweaks
When fapi::Target::set(...) is called to set a Target Handle, the internal
cached ecmd string needs to be deleted because it may no longer be correct
for the new handle. Also some other minor tracing tweaks here that were
requested by the HW team. There is no RTC story for this minor change.

Change-Id: I20c6d99faf31c0d84938fb09af053ae1a698bafd
Reviewed-on: http://gfw160.austin.ibm.com:8080/gerrit/1339
Tested-by: Jenkins Server
Reviewed-by: CAMVAN T. NGUYEN <ctnguyen@us.ibm.com>
Reviewed-by: Brian H. Horton <brianh@linux.ibm.com>
Reviewed-by: A. Patrick Williams III <iawillia@us.ibm.com>