Support modified MSS_CACHE_ENABLE HWPF Attribute
The HW team changed memory_attributes.xml. The target-type of the
MSS_CACHE_ENABLE attribute was changed from system to membuf-chip.
This attribute is initialized by the platform (Hostboot) and
indicates if a Centaur has its L4 cache enabled. The HW team
wanted the ability to turn this off/on for a particular chip.
I also changed from non-volatile/read-only to volatile/writeable
(i.e. on the heap, initialized from PNOR) so that firmware has the
ability to disable an L4 cache, but it is not remembered across a
power cycle. How firmware will decide to actually disable an L4
cache is TBD and this decision may drive further changes to this
attribute (e.g. it could be changed to non-volatile/writeable).
The reason for doing this change now is that the current version
of mss_scominit is accessing MSS_CACHE_ENABLE as a Centaur attribute
and Hostboot fails because it thinks it is a system attribute

Change-Id: Iacd774e482340d95b439579b76230026f38f0440
Reviewed-on: http://gfw160.austin.ibm.com:8080/gerrit/1491
Tested-by: Jenkins Server
Reviewed-by: CAMVAN T. NGUYEN <ctnguyen@us.ibm.com>
Reviewed-by: Van H. Lee <vanlee@us.ibm.com>
Reviewed-by: A. Patrick Williams III <iawillia@us.ibm.com>