HWAS: Apply Partial Good to targets
New HWAS Platform function platReadPartialGood() will be
called by HWAS common function discoverTargets(). The
platReadPartialGood() will query the VPD for the PG keyword
of the CP00 record, and return the vector. The discoverTargets()
function will then parse the record to determine which chiplets
should be marked functional or not, based on the values in the
PG vector.

Change-Id: If29d18a80f8703cc87394cf07d152918957ffd37
RTC: 35078
Reviewed-on: http://gfw160.austin.ibm.com:8080/gerrit/1852
Tested-by: Jenkins Server
Reviewed-by: Daniel M. Crowell <dcrowell@us.ibm.com>
Reviewed-by: Mark W. Wenning <wenning@us.ibm.com>
Reviewed-by: Nicholas E. Bofferding <bofferdn@us.ibm.com>
Reviewed-by: A. Patrick Williams III <iawillia@us.ibm.com>