Fix plugging order for Tuleta in VPD preload script
The original assumptions regarding the DIMM plugging order in
hb-pnor-mvpd-preload.pl was incorrect for Tuleta.  I updated
to be correct.

In a few weeks we should be able to disable this code, but
it's needed for the moment to support the Multi-chip bringup
efforts.

This issue was identified with Defect SW168323.

Change-Id: I3bae0aebbabc9da07f87f379a2f6c2dd15db15c8
Tested-by: Jenkins Server
Reviewed-by: Nicholas E. Bofferding <bofferdn@us.ibm.com>
Reviewed-by: Daniel M. Crowell <dcrowell@us.ibm.com>
Reviewed-by: A. Patrick Williams III <iawillia@us.ibm.com>