Changes so HWSV Loads HB patches on Tuleta in HB Dev Environment
Hardware Server is in the process of enabling function that will
update PNOR with the hostboot base and extended images found
on the FSP during the IPL.  This change triggers a copy command
to be run on the FSP which overwrites the default images
with patch images when in the Hostboot Developer Environment.

Note: This change is backwards compatible, so it does
not need to be co-reqed with an FSP driver.

Change-Id: I5f3dbc49ba66141c63522191721d3e09cd37e6d9
RTC:51082
Reviewed-on: http://gfw160.austin.ibm.com:8080/gerrit/2076
Tested-by: Jenkins Server
Reviewed-by: Daniel M. Crowell <dcrowell@us.ibm.com>
Reviewed-by: Terry J. Opie <opiet@us.ibm.com>
Reviewed-by: A. Patrick Williams III <iawillia@us.ibm.com>