prcd_compile: Support to take a directory as input
Enhance prcd_compile.tcl and prcd_server.tcl scripts to take a -D <dir> option
and build all files (recursively) under that directory. Unknown and duplicate
files will be ignored. Allows the user to point to an eKB sandbox and build a
hostboot image with all hwp code in that sandbox.

Change-Id: I7d99dd0ef94d4aba32db925c672cfdd71a83cfb7
RTC: 35267
Reviewed-on: http://gfw160.austin.ibm.com:8080/gerrit/2084
Tested-by: Jenkins Server
Reviewed-by: Andrew J. Geissler <andrewg@us.ibm.com>
Reviewed-by: A. Patrick Williams III <iawillia@us.ibm.com>