MVPD Accessor Function: setRepairRing
mvpdtest.H comments out tests due to missing mvpd data (issue 51716). Might
not need to be merged depending on the timing of fixing the data.

Brian is working on errltest.H and errluserdetailtest.H to not deconfigure.
As it is mvpd and getRingRepair tests do not run because the dependent
targets are deconfigured. When fixed, then errors occur in mvpdtest. The
problems are commented out for now in fapiwinkletest (get/set Repair Ring).

RTC: 50971
Change-Id: I60d499ebb2d9966ffe2c0cc155a7875820b1ee83
Reviewed-on: http://gfw160.austin.ibm.com:8080/gerrit/2171
Tested-by: Jenkins Server
Reviewed-by: A. Patrick Williams III <iawillia@us.ibm.com>