SW172921 - Bad PNOR chipid in Simics
For some reason the Simics model is returning a different chipid
for the PNOR chip now than it used to.  Talking to Adam we
suspect that the last byte is junk data anyway so it is just old
memory.  The fix is to ignore the last byte since it isn't really
part of the id anyway.

Change-Id: Ie097224c312e5597668b6201eff698f91bfaa950
Reviewed-on: http://gfw160.austin.ibm.com:8080/gerrit/2189
Tested-by: Jenkins Server
Reviewed-by: Andrew J. Geissler <andrewg@us.ibm.com>
Reviewed-by: A. Patrick Williams III <iawillia@us.ibm.com>