HWPF: Add FAPI_SET_SBE_ERROR
When a function extracts an SBE error value, this macro is called which calls
FAPI_SET_HWP_ERROR with the correct error enumerator. This ensures that all
error actions specified in the Error Information XML file are performed. This
macro is generated at build time from the Error Information XML files.

Change-Id: If43dfc7fc2d904408e92d1d640f050e5025bde4b
RTC: 51583
Reviewed-on: http://gfw160.austin.ibm.com:8080/gerrit/2333
Tested-by: Jenkins Server
Reviewed-by: Brian H. Horton <brianh@linux.ibm.com>
Reviewed-by: CAMVAN T. NGUYEN <ctnguyen@us.ibm.com>
Reviewed-by: A. Patrick Williams III <iawillia@us.ibm.com>