Improve PNOR DD Micron NOR Workaround
Improving the workaround implemented in Story 53201 to be more
efficient.  Changing to use the normal SFC polling to determine
when the write/erase operations are complete and then reading
the Micron Flag Status register once to keep the chip from
getting into a bad state.  This should reduce the traffic on the
SPI bus.

Change-Id: I315b165bcd3014a2c3121fd97594e73a2e6c1082
RTC: 61064
Reviewed-on: http://gfw160.austin.ibm.com:8080/gerrit/2795
Tested-by: Jenkins Server
Reviewed-by: Douglas R. Gilbert <dgilbert@us.ibm.com>
Reviewed-by: Daniel M. Crowell <dcrowell@us.ibm.com>
Reviewed-by: A. Patrick Williams III <iawillia@us.ibm.com>