Refactor VPD code to eliminate redundancies
After noticing some redundant code and some odd include gymnastics
I pulled all of the VPD related code (spd,mvpd) into a single vpd
directory/module/component.  This should make the addition of the
centaur fru vpd simpler as well.

Note: this is part of Story 39177 but not all of it, merging this
early to not hold up the work for Story 44009.

Change-Id: I7637a94d22e188050403ed5600b2d7f304c3d006
RTC: 39177
Reviewed-on: http://gfw160.austin.ibm.com:8080/gerrit/2863
Tested-by: Jenkins Server
Reviewed-by: A. Patrick Williams III <iawillia@us.ibm.com>