Add fapi namespace to Chip EC Feature macro
HW Procedure code that queries a Chip EC Feature attribute (to find
out if a chip has a particular feature/bug which is determined by
EC level) will not compile unless the HWP is either within or is
using the fapi namespace. This is because the FAPI generated macros
do not prefix the AttributeID with 'fapi::', the fix is trivial.

Change-Id: Ie65a842d45a9c5387ad3f4d380d9af3507ad817c
Reviewed-on: http://gfw160.austin.ibm.com:8080/gerrit/3681
Reviewed-by: Brian H. Horton <brianh@linux.ibm.com>
Tested-by: Jenkins Server
Reviewed-by: Thi N. Tran <thi@us.ibm.com>
Reviewed-by: A. Patrick Williams III <iawillia@us.ibm.com>