Handle the PR keyword
Add support in the HWAS code to handle the PR keyword, to potentially
force processors to run with less cores than are available.

Move FRU_ID from fsp-only to common, so that the PR keyword code knows
which procs are to be grouped.

This work involves changes to HWAS Common code, as well as hostboot
and FSP platform-specific code.

Change-Id: I6d437a68c37f51edbd6d0d4741f2193892f80a6b
RTC: 35077
Reviewed-on: http://gfw160.austin.ibm.com:8080/gerrit/4221
Tested-by: Jenkins Server
Reviewed-by: A. Patrick Williams III <iawillia@us.ibm.com>