HWPF: Change FAPI_INF traces from TRACS to TRACF
As TRACS traces, we only get FAPI_INF traces by turning them on (which can be
done dynamically). My original intent was that FAPI_INF traces be flight
recorder type traces that may be needed to debug a problem. I suspect that we
will be seeing defects which need FAPI_INF traces to debug and we do not want
to have to recreate and turn on the traces. Patrick agrees to this change.

Change-Id: I29514ef07ab122a8584efb3bb7485f11c81d5bbd
Reviewed-on: http://gfw160.austin.ibm.com:8080/gerrit/4047
Reviewed-by: KUMAR V. CHALLA <kvchalla@us.ibm.com>
Tested-by: Jenkins Server
Reviewed-by: Brian H. Horton <brianh@linux.ibm.com>
Reviewed-by: A. Patrick Williams III <iawillia@us.ibm.com>