HWPF: Allow easy collection of multiple register sets as FFDC on HWP Error
With current code, FAPI Error XML needs to have a separate collectRegisterFfdc
element for each register set, each element specifies the register set and the
target. With this change, you can specify multiple register sets for a
particular target in the same collectRegisterFfdc element. This was requested
by Jeshua and is clearly an improvement.

Change-Id: I234177b63df6da61f2efd579f36336e12a429fbb
Reviewed-on: http://gfw160.austin.ibm.com:8080/gerrit/4213
Tested-by: Jenkins Server
Reviewed-by: Brian H. Horton <brianh@linux.ibm.com>
Reviewed-by: Richard J. Knight <rjknight@us.ibm.com>
Reviewed-by: A. Patrick Williams III <iawillia@us.ibm.com>