ATTN: Pass PRD Centaur instead of processor for channel failure
Currently, HostATTN will pass PRD the Centaur for recoverable and
special attentions, but pass the processor for channel failures reported
by the Centaur. The original reason for the different designs was never
valid in the first place and poorly designed. We are now seeing design
issues in PRD because of the current design so this commit will make
HostATTN pass PRD the Centaur for all attentions types reported by the
Centaur during Hostboot.

Change-Id: I19bd59c4c6efae127b914ff58be171cef1f12fab
Reviewed-on: http://gfw160.aus.stglabs.ibm.com:8080/gerrit/4453
Tested-by: Jenkins Server
Reviewed-by: Sachin Gupta <sgupta2m@in.ibm.com>
Reviewed-by: BENJAMIN J. WEISENBECK <bweisenb@us.ibm.com>
Reviewed-by: Christopher T. Phan <cphan@us.ibm.com>
Reviewed-by: Bilicon Patil <bilpatil@in.ibm.com>
Reviewed-by: A. Patrick Williams III <iawillia@us.ibm.com>