Assign specific values to Targeting TYPEs
Add values to the TYPE enumeration so that the values don't change with inserts
and deletions of types. This will help keep the Targeting Type values to stay
in sync with the HUIDs (the genHwsvMrwXml.pl script that generates HUIDs
hardcodes the type values). The synching mentioned in the issue had already
been done.

Change-Id: I4ee001a825f9a256a3cecb671e3ad1c85c762d54
RTC: 51916
Reviewed-on: http://gfw160.austin.ibm.com:8080/gerrit/4681
Tested-by: Jenkins Server
Reviewed-by: Brian H. Horton <brianh@linux.ibm.com>
Reviewed-by: MIKE J. JONES <mjjones@us.ibm.com>
Reviewed-by: A. Patrick Williams III <iawillia@us.ibm.com>