PRD: fixed issues with secondary registers
Removed secondary type from all MASK registers because we will need to
keep track of which FIR bits PRD has masked.

Removed secondary type from all MBA registers. When PRD is called to do
analysis on an MEMBUF, it will capture all FIRs for the attached MCS and
MBAs in addition to capturing the default registers for the MEMBUF. If
PRD is then called to analyze an attached MBA, PRD will capture all
non-FIR registers on that MBA. However, all previously captured MBA
registers that are marked as secondary will be lost. Therefore, it was
best to remove the secondary type from all MBA registers to avoid the
loss of data.

Change-Id: I41813a538f79b743b49908f8db42369ed2024c00
Reviewed-on: http://gfw160.austin.ibm.com:8080/gerrit/4976
Tested-by: Jenkins Server
Reviewed-by: Christopher T. Phan <cphan@us.ibm.com>
Reviewed-by: Sachin Gupta <sgupta2m@in.ibm.com>
Reviewed-by: A. Patrick Williams III <iawillia@us.ibm.com>
Reviewed-on: http://gfw160.austin.ibm.com:8080/gerrit/5020