Use SBE Setting of Thread Count or Fail
This commit takes out the workaround of using a default setting
of a thread count if the SBE didn't set the right value in a
scratch register.  The current SBE code now does this, and we
will now fail if for some reason the value isn't set.  This commit
also includes sim action file updates to model this behavior.

Change-Id: I83608c402fac675c0287fa3ce38cf75237bcff26
RTC: 63991
Reviewed-on: http://gfw160.austin.ibm.com:8080/gerrit/5255
Tested-by: Jenkins Server
Reviewed-by: A. Patrick Williams III <iawillia@us.ibm.com>