Convert to fixed SLW interfaces
There are two SLW image formats possible, the fixed format is
required to allow concurrent updates to portions of the image.
This is the version that PHYP is using.  To keep things
consistent and tested we need to use the fixed format for the
images produced in Hostboot as well.

Change-Id: I4e1f86b2f5abfcba34d1cdd5f1c2ae9c40c6f02e
RTC: 79613
Reviewed-on: http://gfw160.austin.ibm.com:8080/gerrit/5788
Tested-by: Jenkins Server
Reviewed-by: A. Patrick Williams III <iawillia@us.ibm.com>