Update PNOR Layout XML to indicate HBB has ECC Protection
Just cleaning up the PNOR layout to indicate the HBB partition
has ECC protection.  HWSV is starting to work on support for
injecting ECC in all partitions and will be able to reference
the XML data in all cases to know what to do. (We're not ready
for ECC In the rest of the partitions yet).

Change-Id: I7b9dd365fc4d3feb4c52e1edbd752b3ab4b134ad
Reviewed-on: http://gfw160.austin.ibm.com:8080/gerrit/5794
Tested-by: Jenkins Server
Reviewed-by: A. Patrick Williams III <iawillia@us.ibm.com>