Change HWPF Reg Capture to identify addresses using identifiers
HWP Error XML can specify to capture reigsters when an error happens.
Rather than extracting the address (0xxxx) from the Register XML element
the change here is to use the Register XML element as-is, so it must be
a valid identifier (e.g. from p8_scom_addresses.H) or literal.

Change-Id: I838fd39b088a174cea6ac7d86ca66356ced123fe
RTC: 50833
Reviewed-on: http://gfw160.austin.ibm.com:8080/gerrit/5886
Reviewed-by: Brian H. Horton <brianh@linux.ibm.com>
Tested-by: Jenkins Server
Reviewed-by: Donald E. Dahle <dedahle@us.ibm.com>
Reviewed-by: A. Patrick Williams III <iawillia@us.ibm.com>