Ensure task_crash not optimized out
User space assert() calls the task_crash() macro to crash the thread, this
writes to address NULL, but the write is optimized out. Therefore an assert
does not cause the task to crash as intended. If istep code asserts, the
istepdispatcher does not notice and Hostboot hangs instead of TIing. This
is fixed with a volatile.

Change-Id: Icc27d94d76e937000189655c2f8a640620456673
RTC: 80657
Reviewed-on: http://gfw160.austin.ibm.com:8080/gerrit/5968
Tested-by: Jenkins Server
Reviewed-by: A. Patrick Williams III <iawillia@us.ibm.com>