Merge DDR3/DDR4 attributes with different raw data into common attributes
There are 4 SPD FAPI Attributes where there are separate FAPI Attributes
for DDR3 and DDR4 because the raw data is interpreted differently and
the FAPI enumeration matches the raw data. I agreed with Ken Wright
that for these, there should be a common FAPI attribute that is handled
by an Accessor HWP, the Accessor converts the data appropriately to the
common enumeration. Right now the common enumeration matches DDR3 where
possible to facilitate Cronus switchover from SPD Driver to Accessor,
but eventually this could change. This change helps HWP code to not need
to know the different between DDR3 and DDR4 DIMMs.

Change-Id: I108937ab0bdfc534d9ffc1e391675f4c87875048
RTC: 82631
Reviewed-on: http://gfw160.austin.ibm.com:8080/gerrit/6326
Tested-by: Jenkins Server
Reviewed-by: A. Patrick Williams III <iawillia@us.ibm.com>