Allow Hostboot Debug tools that do unaligned writes to work on Cronus
The ecmd debug framework support uses cipputmempba to write to Hostboot memory,
this was changed some time ago to disallow non cache-line aligned writes. But
some tools (e.g. HwpfAttrOverride) do unaligned writes and they therefore fail.
Also any tool that calls a Hostboot function using CallFunc will fail because
that also does unaligned writes. The emcd team changed cipputmempba to do a
read-modify-write for unaligned writes if the inj mode is specified. Also a
small fix to HwpfAttrOverride here, it was interpreting a decimal value (the
attribute id) as hex which must be due to something changing because this
used to work.

Change-Id: I6535340e41d802489070bc6cffff85b9e33337db
RTC: 70350
Reviewed-on: http://gfw160.austin.ibm.com:8080/gerrit/6362
Reviewed-by: Donald E. Dahle <dedahle@us.ibm.com>
Reviewed-by: Brian H. Horton <brianh@linux.ibm.com>
Tested-by: Jenkins Server
Reviewed-by: A. Patrick Williams III <iawillia@us.ibm.com>