Fix some Targeting attributes to make them Hostboot only
FSP has an FSP-only FSP_PM_SPWUP_OHA_FLAG attribute. There is also a common
PM_SPWUP_OHA_FLAG attribute, this should actually be a HB-only
HB_PM_SPWUP_OHA_FLAG attribute. Also, IS_MPIPL_HB is correctly specified as
a HB-only attribute, but it was incorrectly listed in target_types.xml.
This is part of the fix for SW221921 where FSP was generating duplicate FAPI
attribute access macros for the same FAPI attribute. There is also an FSP
fix to a Targeting makefile for this defect that is done separately.

Change-Id: I0ef46c6bfb46005e0ccad00c9315c6bb46a38da6
CQ: SW221921
Reviewed-on: http://gfw160.austin.ibm.com:8080/gerrit/6453
Tested-by: Jenkins Server
Reviewed-by: Brian H. Horton <brianh@linux.ibm.com>
Reviewed-by: Donald E. Dahle <dedahle@us.ibm.com>
Reviewed-by: A. Patrick Williams III <iawillia@us.ibm.com>