Support new C-VPD FAPI attributes
Old plan was for firmware to switch over some EFF attributes from r/w
storage (Targeting attributes) to VPD (via Accessor), the code was in
place for an easy switch-over. The new plan is to add VPD attribues,
so HWPs decide when to switch-over to use real VPD by changing which
attribute they access (EFF->VPD). This change is to support the new
VPD attributes.

Change-Id: Iac568798a32f6b84bebf79e70b14bb866c6d78fb
RTC: 69935
Tested-by: Jenkins Server
Reviewed-by: Brian H. Horton <brianh@linux.ibm.com>
Reviewed-by: William H. Schwartz <whs@us.ibm.com>
Reviewed-by: A. Patrick Williams III <iawillia@us.ibm.com>