Get PLL ring attribute code needs to error out if EC level not found
The PLL ring Attribute Accessor HWP returns data that depends on the
Chip Name/EC, if the chip EC is not supported, it defaults to providing
data for EC1.0, this caused Thi hours of debug when a new chip (Murano
2.1) was introduced becuase the wrong PLL data was being used, the code
should error out and it would have been obvious that DD2.1 support
needed to be added. Also fix up the error generation and debuggability
in this change.

Change-Id: Idc4ef74fb2613fab1e7e0c17881e6d9112cc6151
CQ: SW241221
RTC: 93108
Reviewed-on: http://gfw160.aus.stglabs.ibm.com:8080/gerrit/8203
Tested-by: Jenkins Server
Reviewed-by: Donald E. Dahle <dedahle@us.ibm.com>
Reviewed-by: Brian H. Horton <brianh@linux.ibm.com>
Reviewed-by: A. Patrick Williams III <iawillia@us.ibm.com>