Retrieve MBA interleave attributes from MRW rather than hardcoding
The MBA interleave values in the MRW are identical to the values hardcoded
by the MRW parser so no concern of breakage. The payload attributes are
hardcoded by the Targeting attribute files, they are not in the MRW and
the values hardcoded by the MRW parser are idendical so safe to just
remove the hardcoding in the MRW parser.

Change-Id: Ie79984588d07a8dad81b54a07b98ec654f7f44a2
RTC: 96087
Reviewed-on: http://gfw160.aus.stglabs.ibm.com:8080/gerrit/8783
Tested-by: Jenkins Server
Reviewed-by: STEPHEN M. CPREK <smcprek@us.ibm.com>
Reviewed-by: A. Patrick Williams III <iawillia@us.ibm.com>