Fix PNOR extended reads and Micron workaround
There are a few issues being addressed here:
1) Added a pollForComplete call in the Micron Flag Status code
since it can take multiple cycles to complete.
2) Added workaround code for Micron bug when reading multi-word
pieces of data, this should fix some bad data we've been getting
for the extended chip id (which is used to enable the above code).
3) Added read of Serial Flash Discovery Parameters to FFDC so we
can have a better idea of which revisions are failing.

Change-Id: Icc983deaac72be3f380269ecc4e321e51ef6005d
CQ: SW246000
Reviewed-on: http://gfw160.aus.stglabs.ibm.com:8080/gerrit/8930
Tested-by: Jenkins Server
Reviewed-by: Michael Baiocchi <baiocchi@us.ibm.com>
Reviewed-by: STEPHEN M. CPREK <smcprek@us.ibm.com>
Reviewed-by: A. Patrick Williams III <iawillia@us.ibm.com>