Remove Initservice Inner Reconfig Loop
PRD enables FIRs in Istep 11. If Istep 12 deconfigures
HW that is asserting FIRs and performs the Inner Reconfig Loop then PRD
ends up logging errors when it sees FIR bits for the deconfigured HW.
The fix is to dispense with the Inner Reconfig Loop, if the code loops
back to Istep 10 then everything is cleaned up. This issue will be
resolved with RTC 101925

Change-Id: I63072e24f18a33bbbfa8c482294e45e020b6cb3f
CQ: SW244527
Backport: release-fips810
Reviewed-on: http://gfw160.aus.stglabs.ibm.com:8080/gerrit/9860
Tested-by: Jenkins Server
Reviewed-by: Andrew J. Geissler <andrewg@us.ibm.com>
Reviewed-by: A. Patrick Williams III <iawillia@us.ibm.com>