SBE Update ensures all processors have Master Processor's SBE Level
A function was added at the very end of the SBE Update procedure to
ensure that any processor continuing with the IPL has the same SBE
Image level as the Master Processor.  Any processor that had an
error during the update process is also deconfigured since we can't
trust its SBE Image level.

Change-Id: Id6dd46ca71ad97ca9f0e6ba30110ea400102a3b4
RTC: 101539
Reviewed-on: http://gfw160.aus.stglabs.ibm.com:8080/gerrit/10868
Tested-by: Jenkins Server
Reviewed-by: Daniel M. Crowell <dcrowell@us.ibm.com>
Reviewed-by: A. Patrick Williams III <iawillia@us.ibm.com>