PRD: In sapphire mode, don't gard target if callout is not primary
In sapphire mode, if there is a strong possibility of a problem
in software and a lower possibility of problem in hardware, we
should not gard the hardware resource. This policy enforcement
intends to prevent the loss of a hardware resource due to a software
problem.

Change-Id: I17c0c3607d17475fc19fa08cd8bedf6f19430485
CQ:SW268025
Backport: release-fips811
Reviewed-on: http://gfw160.aus.stglabs.ibm.com:8080/gerrit/11977
Tested-by: Jenkins Server
Reviewed-by: A. Patrick Williams III <iawillia@us.ibm.com>
Reviewed-by: Zane Shelley <zshelle@us.ibm.com>
Reviewed-on: http://gfw160.aus.stglabs.ibm.com:8080/gerrit/12393